+++ lib/twig/display.rb
@@ -77,6 +77,7 @@ class Twig
       properties = Twig::Branch.all_properties.inject({}) do |result, property_name|
         property = get_branch_property(branch.name, property_name).strip
         property = column(EMPTY_BRANCH_PROPERTY_INDICATOR) if property.empty?
+        property.gsub!(/[\n\r]+/, ' ')
         result.merge(property_name => property)
       end